+++ Mahmoud/game-hub/src/entity/Genre.ts
@@ -1,0 +1,5 @@
+export interface Genre {
+  id: number;
+  name: string;
+  image_background: string;
+}
