+++ Mahmoud/game-hub/src/hooks/useGames.ts
@@ -2,20 +2,9 @@ import { useInfiniteQuery } from "@tanstack/react-query";
 import ms from "ms";
 import APIClient, { FetchResponse } from "../services/api-client";
 import useGamesQueryStore from "../store";
-import { Platform } from "./usePlatforms";
+import { Game } from "../entity/Game";
 
 const apiClient = new APIClient<Game>("/games");
-
-export interface Game {
-  id: number;
-  name: string;
-  slug: string;
-  description_raw: string;
-  background_image: string;
-  parent_platforms: { platform: Platform }[];
-  metacritic: number;
-  rating_top: number;
-}
 
 const useGames = () => {
   const gameQuery = useGamesQueryStore((s) => s.gameQuery);
