+++ Mahmoud/game-hub/src/hooks/useGenres.ts
@@ -1,15 +1,10 @@
 import { useQuery } from "@tanstack/react-query";
 import ms from "ms";
 import genres from "../data/genres";
+import { Genre } from "../entity/Genre";
 import APIClient from "../services/api-client";
 
 const apiClient = new APIClient<Genre>("/genres");
-
-export interface Genre {
-  id: number;
-  name: string;
-  image_background: string;
-}
 
 const useGenres = () =>
   useQuery({
