+++ Mahmoud/game-hub/src/hooks/usePlatforms.ts
@@ -1,15 +1,10 @@
 import { useQuery } from "@tanstack/react-query";
+import ms from "ms";
 import platforms from "../data/platforms";
 import APIClient from "../services/api-client";
-import ms from "ms";
+import { Platform } from "../entity/Platform";
 
 const apiClient = new APIClient<Platform>("/platforms/lists/parents");
-
-export interface Platform {
-  id: number;
-  name: string;
-  slug: string;
-}
 
 const usePlatforms = () =>
   useQuery({
